+++ identity/src/main/java/com/stripe/android/identity/viewmodel/IdentityViewModel.kt
@@ -151,6 +151,26 @@ internal class IdentityViewModel constructor(
     val collectedData: StateFlow<CollectedDataParam> = _collectedData
 
     /**
+     * StateFlow to track request status of postVerificationPageData
+     */
+    @VisibleForTesting
+    internal val verificationPageData = MutableStateFlow<Resource<Unit>>(
+        savedStateHandle[VERIFICATION_PAGE_DATA] ?: run {
+            Resource.idle()
+        }
+    )
+
+    /**
+     * StateFlow to track request status of postVerificationPageSubmit
+     */
+    @VisibleForTesting
+    internal val verificationPageSubmit = MutableStateFlow<Resource<Unit>>(
+        savedStateHandle[VERIFICATION_PAGE_SUBMIT] ?: run {
+            Resource.idle()
+        }
+    )
+
+    /**
      * StateFlow to track missing requirements.
      */
     private val _missingRequirements = MutableStateFlow<List<Requirement>>(
@@ -792,12 +812,18 @@ internal class IdentityViewModel constructor(
     suspend fun postVerificationPageData(
         collectedDataParam: CollectedDataParam
     ): VerificationPageData {
+        verificationPageData.updateStateAndSave {
+            Resource.loading()
+        }
         identityRepository.postVerificationPageData(
             verificationArgs.verificationSessionId,
             verificationArgs.ephemeralKeySecret,
             collectedDataParam,
             calculateClearDataParam(collectedDataParam)
         ).let { verificationPageData ->
+            this.verificationPageData.updateStateAndSave {
+                Resource.success(Unit)
+            }
             _collectedData.updateStateAndSave { oldValue ->
                 oldValue.mergeWith(collectedDataParam)
             }
@@ -824,11 +850,20 @@ internal class IdentityViewModel constructor(
         APIConnectionException::class,
         APIException::class
     )
-    suspend fun postVerificationPageSubmit() =
+    suspend fun postVerificationPageSubmit(): VerificationPageData {
+        verificationPageSubmit.updateStateAndSave {
+            Resource.loading()
+        }
         identityRepository.postVerificationPageSubmit(
             verificationArgs.verificationSessionId,
             verificationArgs.ephemeralKeySecret
-        )
+        ).let {
+            verificationPageSubmit.updateStateAndSave {
+                Resource.success(Unit)
+            }
+            return it
+        }
+    }
 
     fun sendAnalyticsRequest(request: AnalyticsRequestV2) {
         viewModelScope.launch {
@@ -878,6 +913,17 @@ internal class IdentityViewModel constructor(
         }
     }
 
+    /**
+     * Check if there is a outstanding API request being submitted.
+     */
+    fun isSubmitting(): Boolean {
+        return documentFrontUploadedState.value.isLoading() ||
+            documentBackUploadedState.value.isLoading() ||
+            selfieUploadState.value.isAnyLoading() ||
+            verificationPageData.value.status == Status.LOADING ||
+            verificationPageSubmit.value.status == Status.LOADING
+    }
+
     private fun <State> MutableStateFlow<State>.updateStateAndSave(function: (State) -> State) {
         this.update(function)
         savedStateHandle[
@@ -888,6 +934,8 @@ internal class IdentityViewModel constructor(
                 _documentBackUploadedState -> DOCUMENT_BACK_UPLOAD_STATE
                 _collectedData -> COLLECTED_DATA
                 _missingRequirements -> MISSING_REQUIREMENTS
+                verificationPageData -> VERIFICATION_PAGE_DATA
+                verificationPageSubmit -> VERIFICATION_PAGE_SUBMIT
                 else -> {
                     throw IllegalStateException("Unexpected state flow: $this")
                 }
@@ -932,5 +980,7 @@ internal class IdentityViewModel constructor(
         private const val ANALYTICS_STATE = "analytics_upload_state"
         private const val COLLECTED_DATA = "collected_data"
         private const val MISSING_REQUIREMENTS = "missing_requirements"
+        private const val VERIFICATION_PAGE_DATA = "verification_page_data"
+        private const val VERIFICATION_PAGE_SUBMIT = "verification_page_submit"
     }
 }