+++ identity/src/test/java/com/stripe/android/identity/ui/ConfirmationScreenTest.kt
@@ -46,7 +46,9 @@ class ConfirmationScreenTest {
         ) {
             onNodeWithTag(confirmationTitleTag).assertTextEquals(CONFIRMATION_TITLE)
             onNodeWithTag(BODY_TAG).assertTextEquals(CONFIRMATION_BODY)
-            onNodeWithTag(confirmationConfirmButtonTag).assertTextEquals(CONFIRMATION_BUTTON_TEXT)
+            onNodeWithTag(confirmationConfirmButtonTag).assertTextEquals(
+                CONFIRMATION_BUTTON_TEXT.uppercase()
+            )
             onNodeWithTag(confirmationConfirmButtonTag).performClick()
             verify(onConfirmedMock).invoke()
         }