+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/BaseSheetActivity.kt
@@ -146,12 +146,10 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
         viewModel.showLinkVerificationDialog.observe(this) { show ->
             linkAuthView.setContent {
                 if (show) {
-                    viewModel.linkVerificationCallback?.let { callback ->
-                        LinkVerificationDialog(
-                            linkLauncher = viewModel.linkLauncher,
-                            verificationCallback = callback
-                        )
-                    }
+                    LinkVerificationDialog(
+                        linkLauncher = viewModel.linkLauncher,
+                        onResult = viewModel::handleLinkVerificationResult,
+                    )
                 }
             }
         }
@@ -186,12 +184,15 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
         overridePendingTransition(AnimationConstants.FADE_IN, AnimationConstants.FADE_OUT)
     }
 
+    @Deprecated("Deprecated in Java")
     override fun onBackPressed() {
-        if (supportFragmentManager.backStackEntryCount > 0) {
-            clearErrorMessages()
-            super.onBackPressed()
-        } else {
-            viewModel.onUserCancel()
+        if (viewModel.processing.value == false) {
+            if (supportFragmentManager.backStackEntryCount > 0) {
+                clearErrorMessages()
+                super.onBackPressed()
+            } else {
+                viewModel.onUserCancel()
+            }
         }
     }