+++ stripe-core/src/main/java/com/stripe/android/core/utils/Encode.kt
@@ -1,11 +1,13 @@
-package com.stripe.android.stripecardscan.framework.util
+package com.stripe.android.core.utils
 
 import android.util.Base64
+import androidx.annotation.RestrictTo
 import com.stripe.android.core.networking.QueryStringFactory
 import com.stripe.android.core.networking.toMap
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerializationStrategy
 import kotlinx.serialization.json.Json
+import java.net.URLEncoder
 import java.nio.charset.Charset
 
 private val json = Json {
@@ -14,17 +16,20 @@ private val json = Json {
     encodeDefaults = true
 }
 
-internal fun b64Encode(s: String): String =
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun b64Encode(s: String): String =
     Base64.encodeToString(s.toByteArray(Charset.defaultCharset()), Base64.NO_WRAP)
 
-internal fun b64Encode(b: ByteArray): String =
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun b64Encode(b: ByteArray): String =
     Base64.encodeToString(b, Base64.NO_WRAP)
 
 /**
  * Encode a serializable object to a x-www-url-encoded string. The source object must convert to a
  * [Map] so that the parameters can be named.
  */
-internal fun <T> encodeToXWWWFormUrl(serializer: SerializationStrategy<T>, value: T): String =
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun <T> encodeToXWWWFormUrl(serializer: SerializationStrategy<T>, value: T): String =
     QueryStringFactory.createFromParamsWithEmptyValues(
         json.encodeToJsonElement(serializer, value).toMap()
     )
@@ -32,11 +37,19 @@ internal fun <T> encodeToXWWWFormUrl(serializer: SerializationStrategy<T>, value
 /**
  * Encode a serializable object to a JSON string
  */
-internal fun <T> encodeToJson(serializer: SerializationStrategy<T>, value: T): String =
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun <T> encodeToJson(serializer: SerializationStrategy<T>, value: T): String =
     json.encodeToString(serializer, value)
 
 /**
  * Decode an object from a JSON string
  */
-internal fun <T> decodeFromJson(deserializer: DeserializationStrategy<T>, value: String): T =
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun <T> decodeFromJson(deserializer: DeserializationStrategy<T>, value: String): T =
     json.decodeFromString(deserializer, value)
+
+/**
+ * URL-encode a string. This is useful for sanitizing untrusted data for use in URLs.
+ */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun urlEncode(value: String): String = URLEncoder.encode(value, Charsets.UTF_8.name())