+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/payment/ml/CardDetect.kt
@@ -105,8 +105,6 @@ internal class CardDetect private constructor(interpreter: Interpreter) :
         return mlOutput
     }
 
-    override val statsName: String? = null
-
     /**
      * A factory for creating instances of this analyzer.
      */